Split painter_primitive elements
Create painter_primitive.h and move some simple elements
