+++ ImageGenerator.vcxproj
@@ -158,6 +158,7 @@
     <ClInclude Include="ImageGenerator/include/image.h" />
     <ClInclude Include="ImageGenerator/include/image_property.h" />
     <ClInclude Include="ImageGenerator/include/painter.h" />
+    <ClInclude Include="ImageGenerator/include/painter_primitive.h" />
     <ClInclude Include="library/include/stb/stb_image.h" />
     <ClInclude Include="library/include/stb/stb_image_write.h" />
   </ItemGroup>
