+++ ImageGenerator/include/painter.h
@@ -1,50 +1,16 @@
 #pragma once
 
 #include "image.h"
+#include "painter_primitive.h"
 
 #include <array>
 #include <vector>
 
 namespace IMAGE {
 
-struct Coord {
-    double x, y;
-};
-
-struct Color4 {
-    double r, g, b;
-    double a;
-};
-
-struct Color3 {
-    double r, g, b;
-};
-
-using Color = Color4;
-
-struct Circle {
-    Coord center;
-    double radius;
-};
-
-struct Diamond {
-    Coord center;
-    double radius;
-};
-
-struct Rectangle {
-    Coord p1, p2;
-};
-
-struct Triangle {
-    Coord p1, p2, p3;
-};
-
-struct Line {
-    Coord p1, p2;
-    double radius;
-};
-
+/*
+ * Helper to draw primitive elements
+ */
 struct Painter {
     void Draw(Image& image, const Circle& circle, const Color& color) const;
     void Draw(Image& image, const Diamond& diamond, const Color& color) const;
@@ -53,6 +19,8 @@ struct Painter {
     void Draw(Image& image, const Line& line, const Color& color) const;
     void Draw(Image& image, const std::size_t n, const Coord* coords, const Color& color) const;
     void Draw(Image& image, const std::vector<Coord>& coords, const Color& color) const;
+
+    void Draw(Image& image, const Rectangle& rectangle, const Image& sourceImage) const;
 
     template <std::size_t _Sz>
     void Draw(Image& image, const std::array<Coord, _Sz>& coords, const Color& color) const
@@ -67,5 +35,4 @@ struct Painter {
     void DrawGradient(Image& image, const Line& line, const Color3& color1, const Color3& color2) const;
 };
 
-}
-
+} /* namespace IMAGE */
